watchlists/requirements.txt
fastapi[all]==0.81.0
uvicorn[standard]==0.17.6
pytest
psycopg[binary,pool]==3.1.7
jwtdown-fastapi>=0.4.0
python-jose[cryptography]==3.3.0
